requirements.txt
# Optimized requirements for LOW RAM servers (2GB RAM)
# Uses OpenAI API for embeddings instead of local models (no torch/transformers)

# Discord Bot
discord.py>=2.3.2
python-dotenv>=1.0.0

# LLM & AI (core only, no torch/transformers)
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
openai>=1.0.0
langchain-text-splitters>=0.3.0

# Vector Database - PINNED VERSION for compatibility
# v1.9.1 supports named vectors without init_from/strict_mode issues
qdrant-client==1.9.1

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# Data Processing (lightweight)
numpy>=1.24.0
pandas>=2.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Utilities
pyyaml>=6.0.1
pydantic>=2.5.0,<2.10.0
tenacity>=8.2.3
Pillow>=10.0.0

# Web3 & Blockchain
web3>=6.0.0

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.7
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality (optional - remove if not needed)
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.13.0


